+++ src/main/java/in/koreatech/koin/domain/timetable/dto/SemesterResponse.java
@@ -9,11 +9,12 @@ import io.swagger.v3.oas.annotations.media.Schema;
 @JsonNaming(value = SnakeCaseStrategy.class)
 public record SemesterResponse(
     @Schema(description = "id", example = "1")
-    Long id,
+    Integer id,
 
     @Schema(description = "학기", example = "20241")
     String semester
 ) {
+
     public static SemesterResponse from(Semester semester) {
         return new SemesterResponse(
             semester.getId(),
